+++ Classes/Provider/WebPageTypeProvider.php
@@ -10,41 +10,22 @@ namespace Brotkrueml\Schema\Provider;
  * LICENSE.txt file that was distributed with this source code.
  */
 
+use Brotkrueml\Schema\Registry\TypeRegistry;
+use TYPO3\CMS\Core\Utility\GeneralUtility;
+
 /**
  * @internal
  */
 final class WebPageTypeProvider
 {
-    /** @var TypesProvider */
-    private static $typesProvider;
-
     public static function getTypesForTcaSelect(): array
     {
-        $types = static::getTypesProvider()->getWebPageTypes();
+        $types = GeneralUtility::makeInstance(TypeRegistry::class)->getWebPageTypes();
 
         \array_walk($types, function (&$type) {
             $type = [$type, $type];
         });
 
         return \array_merge([['', '']], $types);
-    }
-
-    private static function getTypesProvider(): TypesProvider
-    {
-        if (empty(static::$typesProvider)) {
-            static::$typesProvider = new TypesProvider();
-        }
-
-        return static::$typesProvider;
-    }
-
-    /**
-     * For testing purposes only!
-     *
-     * @param TypesProvider $typesProvider
-     */
-    public static function setTypesProvider(TypesProvider $typesProvider): void
-    {
-        static::$typesProvider = $typesProvider;
     }
 }